+++ docs/release-notes/change-log.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## Unreleased
 
 * Optimized start of output stream in developer mode when start block is in reversible segment and output module does not have any stores in its dependencies.
-
+* Fixed bug where the first streamable block of a chain was not processed correctly when the start block was set to the default zero value.
 
 ## v1.1.19